[Keyboard] annepro2: fix bluetooth multi-pairing issue
This change fixes the bluetooth multi-pairing issue.

** Description **

The bluetooth broadcasting and connection functions in the annepro2_ble.c are incorrect, they do not handle the multi-pairing correctly.

The exiting broadcasting command (12 bytes in total) is the combination of the 11 bytes length machine code (which is stated in the arrary "ble_mcu_start_broadcast[11]") and the 1 byte length value (which is stated in the variable "port" - also known as the Broadcast ID) :

[              ble_mcu_start_broadcast (11 bytes)                 ] + [Broadcast ID 0-3 (1 byte)]
 0x7b, 0x12, 0x53, 0x00, 0x03, 0x00, 0x00, 0x7d, 0x40, 0x01, 0x00,            0x00-0x03

However, with above command, the annepro2 keyboard is always broadcasting the "ID 0" whatever we change the Broadcast ID to any value.

The problem is that the eleventh byte of the command is actually for describing the Broadcast ID, that's why the exiting code is always broadcasting the "ID 0" which makes the keyboard unable to pair with other devices with ID 1-3.

To resolve the problem, we can just simply modify the command to:

[            ble_mcu_start_broadcast (10 bytes)             ] + [Broadcast ID 0-3 (1 byte)] + [ 0x00 (1 byte)]
 0x7b, 0x12, 0x53, 0x00, 0x03, 0x00, 0x00, 0x7d, 0x40, 0x01,            0x00-0x03,                  0x00

The bluetooth connection command has the same issue, so the fix has also covered that issue in this Push Request.